requirements.txt
aiogram==3.0.0b7
python-dotenv
google-api-python-client
gspread
oauth2client